requirements.txt
aiohttp
evaluate
huggingface_hub
jiwer
more_itertools
numba
numpy
openai_whisper==20240930
pandas
praatio
pyaudio
pytorch_lightning==2.5.0.post0
regex
soundfile
tiktoken
tqdm
websockets
wandb
